Crating a functiong for a map and excepting a error
Here I have created a function that returns an integer number multiplied by two. It is called "twice" and a map function will use it for a list, which will be entered by a user. As a result, when a user will start a program the program will ask for a list of numbers. If a user will enter not only integers in the list, then it will ask to rewrite a list again after restarting the program. If everything is correct, then the program will print a list of numbers multiplied by two.