Filter (Solar System)
Here I have created a filter, which return a list of planets, which are in our solar system. A program asks for a list of planets. Then a user has to write a list of names of planets. If it is in solar system, then a program will return this list.