Reduce
To create a reduce function we need to import functools to the code. Also, I created a function ask_numbers_list() to get a list of numbers in any of the higher-ordered functions. Then I created a reduce function. There is a function add() which returns a sum of elements. Reduce function returns a sum of elements that will be entered by a user.