Composing
For composing a useв three of my own functions: add_3 mul_5 and sub_1. They make calculation with the numbers. After that i composed it in one function. And I used it in printing function. A program asks only for a number from a user. If it is not a number, it will return a text with the error text and it will ask to restart a program. If it is a number it will use my three functions and print a result of arythmetical calculations.