Saying (whispering or shouting)
I decided to change the Sorting on whispering and shouting in Saying. A program will ask about the type of speaking either whispering or shouting. Then it asks the text to say to the user. Then there are printed low letters of text if the user chooses whisper type, and capital letters will be written if the user chooses a shouting type of speaking. If there is no one of type, or type would be written incorrectly, then a program will return an error with text and ask to restart itself.